+++ tasks/eslint.js
@@ -15,7 +15,8 @@ module.exports = {
       'tasks/**/*.js',
       'tests/**/*.js',
       'app/**/*.js',
-      'migrations/**/*.js'
+      'migrations/**/*.js',
+      'scripts/**/*.js'
     ]
   }
 };