+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## DEPENDENCIES
 
+- **grunt-spritesmith** has been removed
 - **interactjs** has been upgraded from 1.9.7 to **1.10.11**
 - **node-schedule** has been upgraded from 1.3.2 to **2.0.0**
 - **perfect-scrollbar** has been upgraded from 1.5.0 to **1.5.2**